Code;Docu: Inital commit Udacity Deep Reinforcement Learning Nanodegree Project 2 Continuous Control
The files are copied from the git repo of Udacity deep reinforcement learning nanodegree.
The Unity environments will not be added this repo to keep it as small as possible.
A solution for one and twenty agent soultion will be added in the future.
More details will be added by the next commits.